test(gfql/cypher): polars parity coverage for numeric fns/^/toLower + sign Int64 (review)
Review findings on the =~/numeric-fns branch:
- IMPORTANT: the numeric/^/toLower tests ran only on pandas; parametrize them
  engine=[pandas, polars] so the differential-parity claim is guarded (adds a
  sign() case too). Regression guard for the polars row_pipeline lowerings.
- SUGGESTION: polars sign() returned float vs pandas int; cast to Int64 to match
  the pandas engine and neo4j's Integer-returning spec.

24 parity cases pass (pandas+polars); ruff+mypy clean.

+++ graphistry/compute/gfql/lazy/engine/polars/row_pipeline.py
@@ -114,8 +114,9 @@ def _lower_function(node: FunctionCall, columns: Sequence[str]) -> Optional[pl.E
         # returns float like neo4j/pandas; parity-verified.
         return args[0].cast(pl.Float64).sqrt()
     if name == "sign" and len(args) == 1:
-        # polars .sign() == np.sign for int/float (-1/0/1; null/NaN preserved); parity-verified.
-        return args[0].sign()
+        # polars .sign() == np.sign (-1/0/1; null/NaN preserved); neo4j sign() returns an
+        # Integer, so cast to match the pandas engine (which yields int). Parity-verified.
+        return args[0].sign().cast(pl.Int64)
     if name in {"floor", "ceil", "ceiling"} and len(args) == 1:
         return args[0].ceil() if name in {"ceil", "ceiling"} else args[0].floor()
     if name == "round" and len(args) in {1, 2}:
